Support opened_on date (and Redmine Mentions plugin)

+++ db/migrate/004_add_maillogs_opened.rb
@@ -1,11 +1,12 @@
 class AddMaillogsOpened < ActiveRecord::Migration
 
     def self.up
-        add_column :maillogs, :opened, :boolean, :default => false, :null => false
     end
 
     def self.down
-        remove_column :maillogs, :opened
+        if column_exists?(:maillogs, :opened, :boolean)
+            remove_column :maillogs, :opened
+        end
     end
 
 end
